First version
Implement first version of duplicate-finder.

It builds a list of all the files in the directory tree whose root is
`.`, along with the md5 hash of their contents. It then compares two by
two the hashes and computes the Levenshtein distance between their
names; the distance tolerance is indicated in a variable at the
beginning of the script.

The levenshtein algorithm is not my implementation and is therefore
subject to a particular licensing; all details provided in
levenshtein.py's docstring.